backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
openai==1.58.1
gunicorn==23.0.0
